scheduler/pyproject.toml
[tool.poetry]
name = "scheduler"
version = "0.1.0"
description = ""
authors = ["IamIknow <stepanov.serge97@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiohttp = "^3.8.3"
python-dotenv = "^0.21.0"
aiokafka = "^0.8.0"
redis = "^4.3.5"
psycopg2 = "^2.9.5"
sqlalchemy = "^1.4.44"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
